socialapi: move SessionCache to a separate file
